🔨Fix: 스키마 추가 및 경로 수정
